Moved vocularies to taverna-ro-vocabs
